moved interface to correct package
