+++ app/src/main/java/com/rmkrings/interfaces/IOnBackPressed.java
@@ -1,4 +1,4 @@
-package com.rmkrings.fragments.preferences;
+package com.rmkrings.interfaces;
 
 public interface IOnBackPressed {
